Feat: Implement authentication layout with user state management and routing for protected and auth screens
